Fix sign compare warnings

+++ examples/rasta/c/main.c
@@ -24,7 +24,7 @@ void printHelpAndExit(void){
 }
 
 void addRastaString(struct RastaMessageData * data, int pos, char * str) {
-    int size =  strlen(str) + 1;
+    size_t size = strlen(str) + 1;
 
     struct RastaByteArray msg ;
     allocateRastaByteArray(&msg, size);
