+++ src/rasta/c/config.c
@@ -398,7 +398,7 @@ char* getIpByNic(int index){
 struct RastaIPData extractIPData(char data[256], int arrayIndex) {
     int points = 0;
     int numbers = 0;
-    int pos = 0;
+    size_t pos = 0;
     char port[10];
     struct RastaIPData result;
 
@@ -410,7 +410,7 @@ struct RastaIPData extractIPData(char data[256], int arrayIndex) {
         pos = 1;
     } else {
         //check ip format
-        for (int i = 0; i < strlen(data); i++) {
+        for (size_t i = 0; i < strlen(data); i++) {
             if (isdigit(data[i])) {
                 numbers++;
                 if (numbers > 3) {
@@ -444,7 +444,7 @@ struct RastaIPData extractIPData(char data[256], int arrayIndex) {
 
     //get port
     int j = 0;
-    for (int i = pos+1; i < strlen(data); i++) {
+    for (size_t i = pos+1; i < strlen(data); i++) {
         if (isdigit(data[i])) {
             port[j] = data[i];
         }
@@ -647,7 +647,7 @@ void config_setstd(struct RastaConfig * cfg) {
         cfg->values.redundancy.connections.data = rmalloc(sizeof(struct RastaIPData) * entr.value.array.count);
         cfg->values.redundancy.connections.count = entr.value.array.count;
         //check valid format
-        for (int i = 0; i < entr.value.array.count; i++) {
+        for (unsigned int i = 0; i < entr.value.array.count; i++) {
             struct RastaIPData ip = extractIPData(entr.value.array.data[i].c, i);
             if (ip.port == 0) {
                 logger_log(&cfg->logger,LOG_LEVEL_ERROR, cfg->filename, "RASTA_REDUNDANCY_CONNECTIONS may only contain strings in format ip:port or *:port");
